fix: all tests running
