+++ .github/workflows/s7-plc.yml
@@ -42,5 +42,4 @@ jobs:
       - name: Install Ginkgo
         run: go install github.com/onsi/ginkgo/v2/ginkgo@v2.17.1
       - name: Test
-        working-directory: s7comm_plugin
-        run: TEST_S7_TCPDEVICE=${{ secrets.TEST_S7_TCPDEVICE }} TEST_S7_RACK=${{ secrets.TEST_S7_RACK }} TEST_S7_SLOT=${{ secrets.TEST_S7_SLOT }} make test-s7
+        run: TEST_S7_TCPDEVICE=${{ secrets.TEST_S7_TCPDEVICE }} TEST_S7_RACK=${{ secrets.TEST_S7_RACK }} TEST_S7_SLOT=${{ secrets.TEST_S7_SLOT }} make test
