--- opcua_plugin/opcua_test.go
@@ -1,1123 +1,0 @@
-// Copyright 2023 UMH Systems GmbH
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package opcua_plugin
-
-// NOTE: future tests should be implemented in Ginkgo
-
-import (
-	"context"
-	"encoding/json"
-	"os"
-	"strings"
-	"testing"
-	"time"
-
-	"github.com/gopcua/opcua/ua"
-	"github.com/stretchr/testify/assert"
-)
-
-func TestAgainstSimulator(t *testing.T) {
-	endpoint := os.Getenv("TEST_WAGO_ENDPOINT_URI")
-	username := os.Getenv("TEST_WAGO_USERNAME")
-	password := os.Getenv("TEST_WAGO_PASSWORD")
-
-	// Check if environment variables are set
-	if endpoint != "" || username != "" || password != "" {
-		t.Skip("Skipping test: environment variables are set --> the wago test is running and we are not running a test against the simulator")
-		return
-	}
-
-	t.Run("Connect Subscribe AnalogTypes (simple datatypes)", func(t *testing.T) {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		var err error
-
-		var nodeIDStrings = []string{"ns=6;s=DataAccess_AnalogType_Byte", "ns=6;s=DataAccess_AnalogType_Double", "ns=6;s=DataAccess_AnalogType_Float", "ns=6;s=DataAccess_AnalogType_Int16", "ns=6;s=DataAccess_AnalogType_Int32", "ns=6;s=DataAccess_AnalogType_Int64", "ns=6;s=DataAccess_AnalogType_SByte", "ns=6;s=DataAccess_AnalogType_UInt16", "ns=6;s=DataAccess_AnalogType_UInt32", "ns=6;s=DataAccess_AnalogType_UInt64"}
-
-		parsedNodeIDs := ParseNodeIDs(nodeIDStrings)
-
-		input := &OPCUAInput{
-			Endpoint:         "opc.tcp://localhost:50000",
-			Username:         "",
-			Password:         "",
-			NodeIDs:          parsedNodeIDs,
-			SubscribeEnabled: true,
-		}
-		// Attempt to connect
-		err = input.Connect(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		messageBatch, _, err := input.ReadBatch(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		assert.Equal(t, 10, len(messageBatch))
-
-		for _, message := range messageBatch {
-			message, err := message.AsStructuredMut()
-			if err != nil {
-				t.Fatal(err)
-			}
-			var exampleNumber json.Number = "22.565684"
-			assert.IsType(t, exampleNumber, message) // it should be a number
-			t.Log("Received message: ", message)
-		}
-
-		// Close connection
-		if input.Client != nil {
-			err = input.Client.Close(ctx)
-			if err != nil {
-				t.Fatal(err)
-			}
-		}
-	})
-
-	t.Run("Connect Subscribe null", func(t *testing.T) {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		var err error
-
-		var nodeIDStrings []string = []string{"ns=6;s=DataAccess_DataItem_Null"}
-
-		parsedNodeIDs := ParseNodeIDs(nodeIDStrings)
-
-		input := &OPCUAInput{
-			Endpoint:         "opc.tcp://localhost:50000",
-			Username:         "",
-			Password:         "",
-			NodeIDs:          parsedNodeIDs,
-			SubscribeEnabled: true,
-		}
-		// Attempt to connect
-		err = input.Connect(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		messageBatch, _, err := input.ReadBatch(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		assert.Equal(t, 0, len(messageBatch)) // should never subscribe to null datatype
-
-		// Close connection
-		if input.Client != nil {
-			input.Client.Close(ctx)
-		}
-	})
-
-	t.Run("Connect Subscribe AnalogTypeArray", func(t *testing.T) {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		var err error
-
-		var nodeIDStrings []string = []string{"ns=6;s=DataAccess_AnalogType_Array"}
-
-		parsedNodeIDs := ParseNodeIDs(nodeIDStrings)
-
-		input := &OPCUAInput{
-			Endpoint:         "opc.tcp://localhost:50000",
-			Username:         "",
-			Password:         "",
-			NodeIDs:          parsedNodeIDs,
-			SubscribeEnabled: true,
-		}
-		// Attempt to connect
-		err = input.Connect(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		messageBatch, _, err := input.ReadBatch(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		assert.Equal(t, 26, len(messageBatch)) // Adjust the expected number of items as necessary
-
-		for _, message := range messageBatch {
-			messageParsed, err := message.AsStructuredMut()
-			if err != nil {
-				// This might happen if an empty string is returned from OPC-UA
-				continue
-			}
-
-			opcuapath, found := message.MetaGet("opcua_path")
-			if !found {
-				t.Fatal("Could not find opcua_path")
-			}
-
-			// Determine the data type from the OPC UA path
-			dataType := strings.Split(opcuapath, "_")[5] // This will extract the data type part of the OPC UA path
-			t.Log(dataType)
-
-			// Check if the data type is an array and handle accordingly
-			if strings.HasSuffix(dataType, "Array") {
-				dataTypeOfArray := strings.Split(opcuapath, "_")[6]
-				t.Log(dataTypeOfArray)
-				// Handle array data types
-				switch dataTypeOfArray {
-				case "Duration", "Guid", "LocaleId", "Boolean", "LocalizedText", "NodeId", "QualifiedName", "UtcTime", "DateTime", "Double", "Enumeration", "Float", "Int16", "Int32", "Int64", "Integer", "Number", "SByte", "StatusCode", "String", "UInt16", "UInt32", "UInt64", "UInteger", "Variant", "XmlElement", "ByteString":
-					// Check if the messageParsed is of type slice (array)
-					messageParsedArray, ok := messageParsed.([]interface{})
-					if !ok {
-						t.Errorf("Expected messageParsed to be an array, but got %T: %s : %s", messageParsed, opcuapath, messageParsed)
-					} else {
-						for _, item := range messageParsedArray {
-
-							// Add checking based on the OPC UA path the resulting data type
-							checkDatatypeOfOPCUATag(t, dataTypeOfArray, item, opcuapath)
-						}
-					}
-				case "Byte":
-					checkDatatypeOfOPCUATag(t, "ByteArray", messageParsed, opcuapath)
-				default:
-					t.Errorf("Unsupported array data type in OPC UA path: %s:%s", dataType, opcuapath)
-				}
-			} else {
-				t.Fatalf("Received non-array: %s", opcuapath)
-			}
-		}
-
-		// Close connection
-		if input.Client != nil {
-			input.Client.Close(ctx)
-		}
-	})
-
-	t.Run("Connect Subscribe DataItem", func(t *testing.T) {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		var err error
-
-		var nodeIDStrings = []string{"ns=6;s=DataAccess_DataItem"} // it will subscribe to all values with data type that is non-null.
-
-		parsedNodeIDs := ParseNodeIDs(nodeIDStrings)
-
-		input := &OPCUAInput{
-			Endpoint:         "opc.tcp://localhost:50000",
-			Username:         "",
-			Password:         "",
-			NodeIDs:          parsedNodeIDs,
-			SubscribeEnabled: true,
-		}
-		// Attempt to connect
-		err = input.Connect(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		messageBatch, _, err := input.ReadBatch(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		assert.Equal(t, 23, len(messageBatch)) // these are theoretically >30, but most of them are null, so the browse function ignores them
-
-		for _, message := range messageBatch {
-			messageParsed, err := message.AsStructuredMut()
-			if err != nil {
-				// This might happen if an empty string is returned from OPC-UA
-				continue
-			}
-
-			opcuapath, found := message.MetaGet("opcua_path")
-			if found != true {
-				t.Fatal("Could not find opcua_path")
-			}
-
-			// Determine the data type from the OPC UA path
-			dataType := strings.Split(opcuapath, "_")[5] // This will extract the data type part of the OPC UA path
-
-			// Add checking based on the OPC UA path the resulting data type
-			checkDatatypeOfOPCUATag(t, dataType, messageParsed, opcuapath)
-		}
-
-		// Close connection
-		if input.Client != nil {
-			err = input.Client.Close(ctx)
-			if err != nil {
-				t.Fatal(err)
-			}
-		}
-	})
-
-	t.Run("TestForFailedNodeCrash", func(t *testing.T) {
-		// https://github.com/united-manufacturing-hub/MgmtIssues/issues/1088
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		var err error
-
-		var nodeIDStrings = []string{
-			"ns=3;s=Fast",
-			"ns=3;s=Slow",
-		}
-		parsedNodeIDs := ParseNodeIDs(nodeIDStrings)
-		input := &OPCUAInput{
-			Endpoint:         "opc.tcp://localhost:50000", // Important: ensure that the DNS name in the certificates of the server is also localhost (Hostname and DNS Name), as otherwise the server will refuse the connection
-			Username:         "",
-			Password:         "",
-			NodeIDs:          parsedNodeIDs,
-			SubscribeEnabled: true,
-		}
-
-		// Attempt to connect
-		err = input.Connect(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		messageBatch, _, err := input.ReadBatch(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		assert.NotEmpty(t, messageBatch)
-
-		for _, message := range messageBatch {
-			_, err := message.AsStructured()
-			if err != nil {
-				t.Fatal(err)
-			}
-		}
-
-		// Close connection
-		if input.Client != nil {
-			err = input.Client.Close(ctx)
-			if err != nil {
-				t.Fatal(err)
-			}
-		}
-	})
-
-	t.Run("Connect Subscribe Scalar Arrays", func(t *testing.T) {
-		ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-		defer cancel()
-
-		var err error
-
-		var nodeIDStrings []string = []string{
-			"ns=6;s=Scalar_Static_Arrays_Boolean",
-			"ns=6;s=Scalar_Static_Arrays_Byte",
-			"ns=6;s=Scalar_Static_Arrays_ByteString",
-			"ns=6;s=Scalar_Static_Arrays_DateTime",
-			"ns=6;s=Scalar_Static_Arrays_Double",
-			"ns=6;s=Scalar_Static_Arrays_Duration",
-			"ns=6;s=Scalar_Static_Arrays_Float",
-			"ns=6;s=Scalar_Static_Arrays_Guid",
-			"ns=6;s=Scalar_Static_Arrays_Int16",
-			"ns=6;s=Scalar_Static_Arrays_Int32",
-			"ns=6;s=Scalar_Static_Arrays_Int64",
-			"ns=6;s=Scalar_Static_Arrays_Integer",
-			"ns=6;s=Scalar_Static_Arrays_LocaleId",
-			"ns=6;s=Scalar_Static_Arrays_LocalizedText",
-			"ns=6;s=Scalar_Static_Arrays_NodeId",
-			"ns=6;s=Scalar_Static_Arrays_Number",
-			"ns=6;s=Scalar_Static_Arrays_QualifiedName",
-			"ns=6;s=Scalar_Static_Arrays_SByte",
-			"ns=6;s=Scalar_Static_Arrays_String",
-			"ns=6;s=Scalar_Static_Arrays_UInt16",
-			"ns=6;s=Scalar_Static_Arrays_UInt32",
-			"ns=6;s=Scalar_Static_Arrays_UInt64",
-			"ns=6;s=Scalar_Static_Arrays_UInteger",
-			"ns=6;s=Scalar_Static_Arrays_UtcTime",
-			// "ns=6;s=Scalar_Static_Arrays_Variant", // This node causes a timeout
-			"ns=6;s=Scalar_Static_Arrays_XmlElement",
-		}
-
-		parsedNodeIDs := ParseNodeIDs(nodeIDStrings)
-
-		input := &OPCUAInput{
-			Endpoint:         "opc.tcp://localhost:50000",
-			Username:         "",
-			Password:         "",
-			NodeIDs:          parsedNodeIDs,
-			SubscribeEnabled: true,
-		}
-		// Attempt to connect
-		err = input.Connect(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		messageBatch, _, err := input.ReadBatch(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		assert.Equal(t, 25, len(messageBatch))
-
-		for _, message := range messageBatch {
-			messageParsed, err := message.AsStructuredMut()
-			if err != nil {
-				// This might happen if an empty string is returned from OPC-UA
-				continue
-			}
-
-			opcuapath, found := message.MetaGet("opcua_path")
-			if found != true {
-				t.Fatal("Could not find opcua_path")
-			}
-
-			// Determine the data type from the OPC UA path
-			dataType := strings.Split(opcuapath, "_")[5] // This will extract the data type part of the OPC UA path
-			t.Log(dataType)
-
-			// Check if the data type is an array and handle accordingly
-			if strings.HasSuffix(dataType, "Arrays") {
-				dataTypeOfArray := strings.Split(opcuapath, "_")[6]
-				t.Log(dataTypeOfArray)
-				// Handle array data types
-				switch dataTypeOfArray {
-				case "Duration", "Guid", "LocaleId", "Boolean", "LocalizedText", "NodeId", "QualifiedName", "UtcTime", "DateTime", "Double", "Enumeration", "Float", "Int16", "Int32", "Int64", "SByte", "StatusCode", "String", "UInt16", "UInt32", "UInt64", "XmlElement", "ByteString":
-					// Check if the messageParsed is of type slice (array)
-					messageParsedArray, ok := messageParsed.([]interface{})
-					if !ok {
-						t.Errorf("Expected messageParsed to be an array, but got %T: %s : %s", messageParsed, opcuapath, messageParsed)
-					} else {
-						for _, item := range messageParsedArray {
-
-							// Add checking based on the OPC UA path the resulting data type
-							checkDatatypeOfOPCUATag(t, dataTypeOfArray, item, opcuapath)
-						}
-					}
-				case "Byte":
-					checkDatatypeOfOPCUATag(t, "ByteArray", messageParsed, opcuapath)
-				case "Integer", "Number", "Variant", "UInteger": // Variant Arrays are not supported by go upcua lib
-					t.Logf("Unsupported array data type in OPC UA path: %s:%s", dataType, opcuapath)
-				default:
-					t.Errorf("Unsupported array data type in OPC UA path: %s:%s", dataType, opcuapath)
-				}
-			} else {
-				t.Fatalf("Received non-array: %s", opcuapath)
-			}
-		}
-
-		// Close connection
-		if input.Client != nil {
-			input.Client.Close(ctx)
-		}
-	})
-
-	t.Run("Connect Subscribe does not fail when subscribing to entire simulator", func(t *testing.T) {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		var err error
-
-		var nodeIDStrings []string = []string{"ns=3;s=OpcPlc"}
-
-		parsedNodeIDs := ParseNodeIDs(nodeIDStrings)
-
-		input := &OPCUAInput{
-			Endpoint:         "opc.tcp://localhost:50000",
-			Username:         "",
-			Password:         "",
-			NodeIDs:          parsedNodeIDs,
-			SubscribeEnabled: false, // set to false because some values will change more often in a second resulting in too many messages
-		}
-		// Attempt to connect
-		err = input.Connect(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		messageBatch, _, err := input.ReadBatch(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		assert.GreaterOrEqual(t, len(messageBatch), 125)
-
-		// Close connection
-		if input.Client != nil {
-			input.Client.Close(ctx)
-		}
-	})
-
-	t.Run("Connect does not fail when subscribing to everything", func(t *testing.T) {
-		t.Skip("This might take too long...")
-		ctx, cancel := context.WithTimeout(context.Background(), 60*time.Second)
-		defer cancel()
-
-		var err error
-
-		var nodeIDStrings []string = []string{"i=84"}
-
-		parsedNodeIDs := ParseNodeIDs(nodeIDStrings)
-
-		input := &OPCUAInput{
-			Endpoint:         "opc.tcp://localhost:50000",
-			Username:         "",
-			Password:         "",
-			NodeIDs:          parsedNodeIDs,
-			SubscribeEnabled: true,
-		}
-		// Attempt to connect
-		err = input.Connect(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		messageBatch, _, err := input.ReadBatch(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		assert.Equal(t, 25, len(messageBatch))
-
-		// Close connection
-		if input.Client != nil {
-			input.Client.Close(ctx)
-		}
-	})
-
-	t.Run("Connect Subscribe does subscribe to objects", func(t *testing.T) {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		var err error
-
-		var nodeIDStrings []string = []string{"ns=3;s=SimulatorConfiguration"}
-
-		parsedNodeIDs := ParseNodeIDs(nodeIDStrings)
-
-		input := &OPCUAInput{
-			Endpoint:         "opc.tcp://localhost:50000",
-			Username:         "",
-			Password:         "",
-			NodeIDs:          parsedNodeIDs,
-			SubscribeEnabled: true,
-		}
-		// Attempt to connect
-		err = input.Connect(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// Close connection
-		if input.Client != nil {
-			input.Client.Close(ctx)
-		}
-	})
-
-	t.Run("Connect Subscribe does not fail when subscribing to Anomaly", func(t *testing.T) {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		var err error
-
-		var nodeIDStrings []string = []string{"ns=3;s=Anomaly"}
-
-		parsedNodeIDs := ParseNodeIDs(nodeIDStrings)
-
-		input := &OPCUAInput{
-			Endpoint:         "opc.tcp://localhost:50000",
-			Username:         "",
-			Password:         "",
-			NodeIDs:          parsedNodeIDs,
-			SubscribeEnabled: false, // disabling subscribe because messages change moreo ften than once in a second reuslting in to many messages
-		}
-		// Attempt to connect
-		err = input.Connect(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		messageBatch, _, err := input.ReadBatch(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		assert.Equal(t, 4, len(messageBatch))
-
-		// Close connection
-		if input.Client != nil {
-			input.Client.Close(ctx)
-		}
-	})
-
-	t.Run("Connect Subscribe does not fail when subscribing to Basic", func(t *testing.T) {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		var err error
-
-		var nodeIDStrings []string = []string{"ns=3;s=Basic"}
-
-		parsedNodeIDs := ParseNodeIDs(nodeIDStrings)
-
-		input := &OPCUAInput{
-			Endpoint:         "opc.tcp://localhost:50000",
-			Username:         "",
-			Password:         "",
-			NodeIDs:          parsedNodeIDs,
-			SubscribeEnabled: false, // disabling subscribe because messages change moreo ften than once in a second reuslting in to many messages
-		}
-		// Attempt to connect
-		err = input.Connect(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		messageBatch, _, err := input.ReadBatch(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		assert.Equal(t, 4, len(messageBatch))
-
-		// Close connection
-		if input.Client != nil {
-			input.Client.Close(ctx)
-		}
-	})
-
-	t.Run("Connect Subscribe does not fail when subscribing to Deterministic GUID", func(t *testing.T) {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		var err error
-
-		var nodeIDStrings []string = []string{"ns=3;s=Deterministic GUID"}
-
-		parsedNodeIDs := ParseNodeIDs(nodeIDStrings)
-
-		input := &OPCUAInput{
-			Endpoint:         "opc.tcp://localhost:50000",
-			Username:         "",
-			Password:         "",
-			NodeIDs:          parsedNodeIDs,
-			SubscribeEnabled: false,
-		}
-		// Attempt to connect
-		err = input.Connect(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		messageBatch, _, err := input.ReadBatch(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		assert.Equal(t, 5, len(messageBatch))
-
-		// Close connection
-		if input.Client != nil {
-			input.Client.Close(ctx)
-		}
-	})
-
-	t.Run("Connect Subscribe does not fail when subscribing to Fast", func(t *testing.T) {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		var err error
-
-		var nodeIDStrings []string = []string{"ns=3;s=Fast"}
-
-		parsedNodeIDs := ParseNodeIDs(nodeIDStrings)
-
-		input := &OPCUAInput{
-			Endpoint:         "opc.tcp://localhost:50000",
-			Username:         "",
-			Password:         "",
-			NodeIDs:          parsedNodeIDs,
-			SubscribeEnabled: false,
-		}
-		// Attempt to connect
-		err = input.Connect(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		messageBatch, _, err := input.ReadBatch(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		assert.GreaterOrEqual(t, len(messageBatch), 5)
-
-		// Close connection
-		if input.Client != nil {
-			input.Client.Close(ctx)
-		}
-	})
-
-	t.Run("Connect Subscribe does not fail when subscribing to Slow", func(t *testing.T) {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		var err error
-
-		var nodeIDStrings []string = []string{"ns=3;s=Slow"}
-
-		parsedNodeIDs := ParseNodeIDs(nodeIDStrings)
-
-		input := &OPCUAInput{
-			Endpoint:         "opc.tcp://localhost:50000",
-			Username:         "",
-			Password:         "",
-			NodeIDs:          parsedNodeIDs,
-			SubscribeEnabled: false,
-		}
-		// Attempt to connect
-		err = input.Connect(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		messageBatch, _, err := input.ReadBatch(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		assert.GreaterOrEqual(t, len(messageBatch), 100)
-
-		// Close connection
-		if input.Client != nil {
-			input.Client.Close(ctx)
-		}
-	})
-
-	t.Run("Connect Subscribe does not fail when subscribing to Special", func(t *testing.T) {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		var err error
-
-		var nodeIDStrings []string = []string{"ns=3;s=Special"}
-
-		parsedNodeIDs := ParseNodeIDs(nodeIDStrings)
-
-		input := &OPCUAInput{
-			Endpoint:         "opc.tcp://localhost:50000",
-			Username:         "",
-			Password:         "",
-			NodeIDs:          parsedNodeIDs,
-			SubscribeEnabled: false,
-		}
-		// Attempt to connect
-		err = input.Connect(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		messageBatch, _, err := input.ReadBatch(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		assert.Equal(t, 7, len(messageBatch))
-
-		// Close connection
-		if input.Client != nil {
-			input.Client.Close(ctx)
-		}
-	})
-
-}
-
-func checkDatatypeOfOPCUATag(t *testing.T, dataType string, messageParsed any, opcuapath string) {
-	t.Logf("%s, %+v, %s", dataType, messageParsed, opcuapath)
-	switch dataType {
-	case "Boolean":
-		var expectedType bool
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received Boolean message: ", messageParsed)
-
-	case "Byte":
-		var expectedType json.Number
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received Byte message: ", messageParsed)
-
-	case "DateTime": //warning: there is a bug when the date is a lot in the future year 30828
-		var expectedType string
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received DateTime message: ", messageParsed)
-
-	case "Double":
-		var expectedType json.Number
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received Double message: ", messageParsed)
-
-	case "Enumeration":
-		var expectedType json.Number
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received Enumeration message: ", messageParsed)
-	case "ExpandedNodeId":
-		// Assert that messageParsed is of type map[string]interface{}
-		parsedMap, ok := messageParsed.(map[string]interface{})
-		if !ok {
-			t.Errorf("Expected messageParsed to be of type map[string]interface{}, but got %T", messageParsed)
-		} else {
-			// Check if the keys inside the map are correct
-			expectedKeys := []string{"NamespaceURI", "NodeID", "ServerIndex"}
-			for _, key := range expectedKeys {
-				if _, exists := parsedMap[key]; !exists {
-					t.Errorf("Expected key %s missing in messageParsed", key)
-				}
-			}
-
-			// Optionally, log the received message
-			t.Log("Received ExpandedNodeId message: ", messageParsed)
-		}
-	case "Float":
-		var expectedType json.Number
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received Float message: ", messageParsed)
-
-	case "Guid":
-		// Assert that messageParsed is of type map[string]interface{}
-		parsedMap, ok := messageParsed.(map[string]interface{})
-		if !ok {
-			t.Errorf("Expected messageParsed to be of type map[string]interface{}, but got %T", messageParsed)
-		} else {
-			// Check if the keys inside the map are correct
-			expectedKeys := []string{"Data1", "Data2", "Data3", "Data4"}
-			for _, key := range expectedKeys {
-				if _, exists := parsedMap[key]; !exists {
-					t.Errorf("Expected key %s missing in messageParsed", key)
-				}
-			}
-
-			// Optionally, log the received message
-			t.Log("Received GUID message: ", messageParsed)
-		}
-	case "Int16":
-		var expectedType json.Number
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received Int16 message: ", messageParsed)
-
-	case "Int32":
-		var expectedType json.Number
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received Int32 message: ", messageParsed)
-
-	case "Int64":
-		var expectedType json.Number
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received Int64 message: ", messageParsed)
-
-	case "Integer":
-		var expectedType json.Number
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received Integer message: ", messageParsed)
-
-	case "LocalizedText":
-		// Assert that messageParsed is of type map[string]interface{}
-		parsedMap, ok := messageParsed.(map[string]interface{})
-		if !ok {
-			t.Errorf("Expected messageParsed to be of type map[string]interface{}, but got %T", messageParsed)
-		} else {
-			// Check if the keys inside the map are correct
-			expectedKeys := []string{"EncodingMask", "Locale", "Text"}
-			for _, key := range expectedKeys {
-				if _, exists := parsedMap[key]; !exists {
-					t.Errorf("Expected key %s missing in messageParsed", key)
-				}
-			}
-
-			// Optionally, log the received message
-			t.Log("Received LocalizedText message: ", messageParsed)
-		}
-	case "NodeId":
-		var expectedType string
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received NodeId message: ", messageParsed)
-	case "Number":
-		var expectedType json.Number
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received Number message: ", messageParsed)
-	case "QualifiedName":
-		// Assert that messageParsed is of type map[string]interface{}
-		parsedMap, ok := messageParsed.(map[string]interface{})
-		if !ok {
-			t.Errorf("Expected messageParsed to be of type map[string]interface{}, but got %T", messageParsed)
-		} else {
-			// Define the keys expected in a QualifiedName message
-			expectedKeys := []string{"NamespaceIndex", "Name"}
-			for _, key := range expectedKeys {
-				if _, exists := parsedMap[key]; !exists {
-					t.Errorf("Expected key %s missing in messageParsed", key)
-				}
-			}
-
-			// Optionally, log the received message
-			t.Log("Received QualifiedName message: ", messageParsed)
-		}
-	case "SByte":
-		var expectedType json.Number
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received SByte message: ", messageParsed)
-	case "StatusCode":
-		var expectedType json.Number
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received StatusCode message: ", messageParsed)
-	case "String":
-		var expectedType string
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received String message: ", messageParsed)
-	case "UInt16":
-		var expectedType json.Number
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received UInt16 message: ", messageParsed)
-	case "UInt32":
-		var expectedType json.Number
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received UInt32 message: ", messageParsed)
-	case "UInt64":
-		var expectedType json.Number
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received UInt64 message: ", messageParsed)
-	case "UInteger":
-		var expectedType json.Number
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received UInteger message: ", messageParsed)
-	case "ByteArray": // case as an array of bytes is not a number, but a string
-		var expectedType string
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received ByteArray message: ", messageParsed)
-	case "ByteString":
-		var expectedType string
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received ByteString message: ", messageParsed)
-	case "Duration":
-		var expectedType json.Number
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received Duration message: ", messageParsed)
-	case "LocaleId":
-		var expectedType string
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received LocaleId message: ", messageParsed)
-	case "UtcTime":
-		var expectedType string
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received UtcTime message: ", messageParsed)
-	case "Variant":
-		var expectedType map[string]interface{}
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received Variant message: ", messageParsed)
-	case "XmlElement":
-		var expectedType string
-		assert.IsType(t, expectedType, messageParsed)
-		t.Log("Received XmlElement message: ", messageParsed)
-	default:
-		t.Errorf("Unsupported data type in OPC UA path: %s:%s", dataType, opcuapath)
-	}
-}
-
-func MockGetEndpoints() []*ua.EndpointDescription {
-	// Define the mock endpoints with the desired properties
-	endpoint1 := &ua.EndpointDescription{
-		EndpointURL: "opc.tcp://example.com:4840", // Replace with your actual server URL
-		Server: &ua.ApplicationDescription{
-			ApplicationURI:  "urn:example:server", // Replace with your server's URI
-			ApplicationType: ua.ApplicationTypeServer,
-		},
-		ServerCertificate: []byte{},                                                    // Replace with your server certificate
-		SecurityMode:      ua.MessageSecurityModeFromString("SignAndEncrypt"),          // Use appropriate security mode
-		SecurityPolicyURI: "http://opcfoundation.org/UA/SecurityPolicy#Basic256Sha256", // Use appropriate security policy URI
-		UserIdentityTokens: []*ua.UserTokenPolicy{
-			{
-				PolicyID:          "anonymous",
-				TokenType:         ua.UserTokenTypeAnonymous,
-				IssuedTokenType:   "http://opcfoundation.org/UA/UserTokenPolicy#Anonymous",
-				SecurityPolicyURI: "http://opcfoundation.org/UA/SecurityPolicy#Basic256Sha256",
-			},
-			{
-				PolicyID:          "username",
-				TokenType:         ua.UserTokenTypeUserName,
-				IssuedTokenType:   "http://opcfoundation.org/UA/UserTokenPolicy#UserName",
-				SecurityPolicyURI: "http://opcfoundation.org/UA/SecurityPolicy#Basic256Sha256",
-			},
-		},
-		TransportProfileURI: "http://opcfoundation.org/UA-Profile/Transport/uatcp-uasc-uabinary",
-		SecurityLevel:       3, // Use an appropriate security level
-	}
-
-	endpoint2 := &ua.EndpointDescription{
-		EndpointURL: "opc.tcp://example2.com:4840", // Replace with your actual server URL
-		Server: &ua.ApplicationDescription{
-			ApplicationURI:  "urn:example2:server", // Replace with your server's URI
-			ApplicationType: ua.ApplicationTypeServer,
-		},
-		ServerCertificate: []byte("mock_certificate_2"),                      // Replace with your server certificate
-		SecurityMode:      ua.MessageSecurityModeFromString("None"),          // Use appropriate security mode
-		SecurityPolicyURI: "http://opcfoundation.org/UA/SecurityPolicy#None", // Use appropriate security policy URI
-		UserIdentityTokens: []*ua.UserTokenPolicy{
-			{
-				PolicyID:          "anonymous",
-				TokenType:         ua.UserTokenTypeAnonymous,
-				IssuedTokenType:   "http://opcfoundation.org/UA/UserTokenPolicy#Anonymous",
-				SecurityPolicyURI: "http://opcfoundation.org/UA/SecurityPolicy#None",
-			},
-			{
-				PolicyID:          "username",
-				TokenType:         ua.UserTokenTypeUserName,
-				IssuedTokenType:   "http://opcfoundation.org/UA/UserTokenPolicy#UserName",
-				SecurityPolicyURI: "http://opcfoundation.org/UA/SecurityPolicy#None",
-			},
-		},
-		TransportProfileURI: "http://opcfoundation.org/UA-Profile/Transport/uatcp-uasc-uabinary",
-		SecurityLevel:       0, // Use an appropriate security level
-	}
-
-	endpoint3 := &ua.EndpointDescription{
-		EndpointURL: "opc.tcp://example3.com:4840", // Replace with your actual server URL
-		Server: &ua.ApplicationDescription{
-			ApplicationURI:  "urn:example3:server", // Replace with your server's URI
-			ApplicationType: ua.ApplicationTypeServer,
-		},
-		ServerCertificate: []byte("mock_certificate_2"),                                    // Replace with your server certificate
-		SecurityMode:      ua.MessageSecurityModeFromString("SignAndEncrypt"),              // Use appropriate security mode
-		SecurityPolicyURI: "http://opcfoundation.org/UA/SecurityPolicy#Aes256Sha256RsaPss", // Use appropriate security policy URI
-		UserIdentityTokens: []*ua.UserTokenPolicy{
-			{
-				PolicyID:          "anonymous",
-				TokenType:         ua.UserTokenTypeAnonymous,
-				IssuedTokenType:   "http://opcfoundation.org/UA/UserTokenPolicy#Anonymous",
-				SecurityPolicyURI: "http://opcfoundation.org/UA/SecurityPolicy#Aes256Sha256RsaPss",
-			},
-			{
-				PolicyID:          "username",
-				TokenType:         ua.UserTokenTypeUserName,
-				IssuedTokenType:   "http://opcfoundation.org/UA/UserTokenPolicy#UserName",
-				SecurityPolicyURI: "http://opcfoundation.org/UA/SecurityPolicy#Aes256Sha256RsaPss",
-			},
-		},
-		TransportProfileURI: "http://opcfoundation.org/UA-Profile/Transport/uatcp-uasc-uabinary",
-		SecurityLevel:       0, // Use an appropriate security level
-	}
-
-	// Return the mock endpoints as a slice
-	return []*ua.EndpointDescription{endpoint1, endpoint2, endpoint3}
-}
-
-func TestGetReasonableEndpoint_Insecure(t *testing.T) {
-	input := &OPCUAInput{
-		Endpoint: "",
-		Username: "",
-		Password: "",
-		NodeIDs:  nil,
-		Insecure: true,
-	}
-
-	endpoints := MockGetEndpoints()
-	selectedEndpoint := input.getReasonableEndpoint(endpoints, ua.UserTokenTypeFromString("Anonymous"), input.Insecure, "", "")
-
-	if selectedEndpoint != nil {
-		if selectedEndpoint.SecurityMode != ua.MessageSecurityModeFromString("None") {
-			t.Errorf("Expected selected endpoint to have no encryption, but got %v", selectedEndpoint.SecurityMode)
-		}
-	} else {
-		t.Fatalf("Expected a reasonable endpoint, but got nil") // This needs to be fatal, to prevent nil error in selectedEndpoint2 check
-	}
-
-	input2 := &OPCUAInput{
-		Endpoint: "",
-		Username: "",
-		Password: "",
-		NodeIDs:  nil,
-		Insecure: false,
-	}
-
-	selectedEndpoint2 := input.getReasonableEndpoint(endpoints, ua.UserTokenTypeFromString("Anonymous"), input2.Insecure, "", "")
-
-	if selectedEndpoint2 != nil {
-		if selectedEndpoint2.SecurityMode != ua.MessageSecurityModeFromString("SignAndEncrypt") {
-			t.Errorf("Expected selected endpoint to have encryption, but got %v", selectedEndpoint.SecurityMode)
-		}
-	} else {
-		t.Error("Expected a reasonable endpoint, but got nil")
-	}
-}
-
-func TestGetReasonableEndpoint_SecurityModeAndPolicy(t *testing.T) {
-	input := &OPCUAInput{
-		Endpoint:       "",
-		Username:       "123",
-		Password:       "213",
-		NodeIDs:        nil,
-		Insecure:       false,
-		SecurityMode:   "SignAndEncrypt",
-		SecurityPolicy: "Aes256Sha256RsaPss",
-	}
-
-	endpoints := MockGetEndpoints()
-	selectedEndpoint := input.getReasonableEndpoint(endpoints, ua.UserTokenTypeFromString("UserName"), input.Insecure, input.SecurityMode, input.SecurityPolicy)
-
-	if selectedEndpoint != nil {
-		if selectedEndpoint.SecurityMode != ua.MessageSecurityModeFromString(input.SecurityMode) && selectedEndpoint.SecurityPolicyURI != "http://opcfoundation.org/UA/SecurityPolicy#"+input.SecurityPolicy {
-			t.Errorf("Expected selected endpoint to have encryption with security mode %v and policy %v, but got %v and %v", input.SecurityMode, input.SecurityPolicy, selectedEndpoint.SecurityMode, selectedEndpoint.SecurityPolicyURI)
-		}
-	} else {
-		t.Error("Expected a reasonable endpoint, but got nil")
-	}
-}
-
-func TestUpdateNodePaths(t *testing.T) {
-	tests := []struct {
-		name     string
-		nodes    []NodeDef
-		expected []NodeDef
-	}{
-		{
-			name: "no duplicates",
-			nodes: []NodeDef{
-				{Path: "Folder.Tag1", NodeID: ua.MustParseNodeID("ns=1;s=node1")},
-				{Path: "Folder.Tag2", NodeID: ua.MustParseNodeID("ns=1;s=node2")},
-				{Path: "Folder.Tag3", NodeID: ua.MustParseNodeID("ns=1;s=node3")},
-			},
-			expected: []NodeDef{
-				{Path: "Folder.Tag1", NodeID: ua.MustParseNodeID("ns=1;s=node1")},
-				{Path: "Folder.Tag2", NodeID: ua.MustParseNodeID("ns=1;s=node2")},
-				{Path: "Folder.Tag3", NodeID: ua.MustParseNodeID("ns=1;s=node3")},
-			},
-		},
-		{
-			name: "duplicates",
-			nodes: []NodeDef{
-				{Path: "Folder.Tag1", NodeID: ua.MustParseNodeID("ns=1;s=node1")},
-				{Path: "Folder.Tag1", NodeID: ua.MustParseNodeID("ns=1;s=node2")},
-				{Path: "Folder.Tag2", NodeID: ua.MustParseNodeID("ns=1;s=node3")},
-				{Path: "Folder.Tag3", NodeID: ua.MustParseNodeID("ns=1;s=node4")},
-			},
-			expected: []NodeDef{
-				{Path: "Folder.ns_1_s_node1", NodeID: ua.MustParseNodeID("ns=1;s=node1")},
-				{Path: "Folder.ns_1_s_node2", NodeID: ua.MustParseNodeID("ns=1;s=node2")},
-				{Path: "Folder.Tag2", NodeID: ua.MustParseNodeID("ns=1;s=node3")},
-				{Path: "Folder.Tag3", NodeID: ua.MustParseNodeID("ns=1;s=node4")},
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			updateNodePaths(tt.nodes)
-			assert.Equal(t, tt.expected, tt.nodes)
-		})
-	}
-}
